Implement Entity Processing - InProgress #22
Revise how entities process their messages. The flow goes like this:
Controller asks Entity for the command for a particular set of arguments.
Entity returns a subclass of the appropriate kind of command whith the
apply method overridden to do the actual work (default is a Noop)
Controller passes that command to Scrupal.handle which uses actor.ask
to forward it to the dispatcher which queues it for execution in another
thread. Meanwhile Scrupal gets back a Future[Result[_]] which it
puts into the completion of the request. The dispatcher is the
small actor in EntityProcessor that just guards the entity's
command's apply function from throwing exceptions. In that case an
ExceptionResult is returned instead of one of the types actually
carrying a payload.

Added NodeProcessor as a separate processor for nodes but it might just
end up working the same way as EntityProcessor. All other changes are
related to this change. Scrupal runs after this change but it fails
on routing in simple tests. I'll fix that this weekend.