Generalize Routing A Little More - InProgress #33 * Make a distinction between terminal ActionProviders and those that have subordinates * Make Scrupal.load overridable, override in CoreScrupal to initialize WelcomeSite * Simplify controller construction: just Assets & ActionProvider * Fix rawPathPrefixWithMatch to match the longest paths first (match echoes before echo) * Revise ActionProviderController to adapt to whatever active hierarchy is available once a site is found
The basic idea here is that ActionProviders can either consume one segment of the path
and defer to subordinates or they can consume the whole path and provide an action (terminal case).
This arrangement means we have one controller for all ActionProviders and it is always "fresh" as
it adapts to enablements since it always asks each provider for its list of subordinates.